+++ lightning_pose/data/utils.py
@@ -9,7 +9,13 @@ import torch
 from torchtyping import TensorType
 from typeguard import typechecked
 
-from lightning_pose.data.datatypes import SemiSupervisedDataLoaderDict
+from lightning_pose.data.datatypes import (
+    HeatmapLabeledBatchDict,
+    MultiviewHeatmapLabeledBatchDict,
+    MultiviewUnlabeledBatchDict,
+    SemiSupervisedDataLoaderDict,
+    UnlabeledBatchDict,
+)
 
 # to ignore imports for sphix-autoapidoc
 __all__ = [
@@ -22,6 +28,8 @@ __all__ = [
     "evaluate_heatmaps_at_location",
     "undo_affine_transform",
     "undo_affine_transform_batch",
+    "normalized_to_bbox",
+    "convert_bbox_coords",
 ]
 
 
@@ -529,3 +537,66 @@ def undo_affine_transform_batch(
         keypoints_unaugmented = keypoints_augmented
 
     return keypoints_unaugmented
+
+
+def normalized_to_bbox(
+    keypoints: TensorType["batch", "num_keypoints", "xy":2],
+    bbox: TensorType["batch", "xyhw":4]
+) -> TensorType["batch", "num_keypoints", "xy":2]:
+    if keypoints.shape[0] == bbox.shape[0]:
+        # normal batch
+        keypoints[:, :, 0] *= bbox[:, 3].unsqueeze(1)  # scale x by box width
+        keypoints[:, :, 0] += bbox[:, 0].unsqueeze(1)  # add bbox x offset
+        keypoints[:, :, 1] *= bbox[:, 2].unsqueeze(1)  # scale y by box height
+        keypoints[:, :, 1] += bbox[:, 1].unsqueeze(1)  # add bbox y offset
+    else:
+        # context batch; we don't have predictions for first/last two frames
+        keypoints[:, :, 0] *= bbox[2:-2, 3].unsqueeze(1)  # scale x by box width
+        keypoints[:, :, 0] += bbox[2:-2, 0].unsqueeze(1)  # add bbox x offset
+        keypoints[:, :, 1] *= bbox[2:-2, 2].unsqueeze(1)  # scale y by box height
+        keypoints[:, :, 1] += bbox[2:-2, 1].unsqueeze(1)  # add bbox y offset
+    return keypoints
+
+
+def convert_bbox_coords(
+    batch_dict: (
+        HeatmapLabeledBatchDict
+        | MultiviewHeatmapLabeledBatchDict
+        | MultiviewUnlabeledBatchDict
+        | UnlabeledBatchDict
+    ),
+    predicted_keypoints: TensorType["batch", "num_targets"],
+) -> TensorType["batch", "num_targets"]:
+    """Transform keypoints from bbox coordinates to absolute frame coordinates."""
+    num_targets = predicted_keypoints.shape[1]
+    num_keypoints = num_targets // 2
+    # reshape from (batch, n_targets) back to (batch, n_key, 2), in x,y order
+    predicted_keypoints = predicted_keypoints.reshape((-1, num_keypoints, 2))
+    # divide by image dims to get 0-1 normalized coordinates
+    if "images" in batch_dict.keys():
+        predicted_keypoints[:, :, 0] /= batch_dict["images"].shape[-1]  # -1 dim is width "x"
+        predicted_keypoints[:, :, 1] /= batch_dict["images"].shape[-2]  # -2 dim is height "y"
+    else:  # we have unlabeled dict, 'frames' instead of 'images'
+        predicted_keypoints[:, :, 0] /= batch_dict["frames"].shape[-1]  # -1 dim is width "x"
+        predicted_keypoints[:, :, 1] /= batch_dict["frames"].shape[-2]  # -2 dim is height "y"
+    # multiply and add by bbox dims (x,y,h,w)
+    if "num_views" in batch_dict.keys() and int(batch_dict["num_views"].max()) > 1:
+        unique = batch_dict["num_views"].unique()
+        if len(unique) != 1:
+            raise ValueError(
+                f"each batch element must contain the same number of views; "
+                f"found elements with {unique} views"
+            )
+        num_views = int(unique)
+        num_keypoints_per_view = num_keypoints // num_views
+        for v in range(num_views):
+            idx_beg = num_keypoints_per_view * v
+            idx_end = num_keypoints_per_view * (v + 1)
+            predicted_keypoints[:, idx_beg:idx_end, :] = normalized_to_bbox(
+                predicted_keypoints[:, idx_beg:idx_end, :],
+                batch_dict["bbox"][:, 4 * v:4 * (v + 1)],
+            )
+    else:
+        predicted_keypoints = normalized_to_bbox(predicted_keypoints, batch_dict["bbox"])
+    # return new keypoints, reshaped to (batch, num_targets)
+    return predicted_keypoints.reshape((-1, num_targets))
