+++ lightning_pose/models/heatmap_tracker_mhcrnn.py
@@ -14,14 +14,13 @@ from lightning_pose.data.datatypes import (
     MultiviewUnlabeledBatchDict,
     UnlabeledBatchDict,
 )
-from lightning_pose.data.utils import undo_affine_transform_batch
+from lightning_pose.data.utils import convert_bbox_coords, undo_affine_transform_batch
 from lightning_pose.losses.factory import LossFactory
 from lightning_pose.losses.losses import RegressionRMSELoss
 from lightning_pose.models.backbones import ALLOWED_BACKBONES
 from lightning_pose.models.base import (
     BaseSupervisedTracker,
     SemiSupervisedTrackerMixin,
-    convert_bbox_coords,
 )
 from lightning_pose.models.heads import HeatmapMHCRNNHead
