+++ tests/data/test_utils.py
@@ -2,6 +2,7 @@
 
 import copy
 
+import numpy as np
 import pytest
 import torch
 from kornia.geometry.subpix import spatial_expectation2d, spatial_softmax2d
@@ -561,48 +562,142 @@ def test_undo_affine_transform_batch():
     assert torch.allclose(keypoints, keypoints_noaug, atol=1e-4)
 
 
-# def test_heatmap_generation():
-#
-#     # want to compare the output of our manual function to kornia's
-#     # if it works, move to kornia
-#
-#     # a batch size of 2, with 3 keypoints per batch.
-#     from time import time
-#     from kornia.geometry.subpix import render_gaussian2d
-#
-#     batch_dim_1 = torch.tensor([[1.0, 2.0], [3.0, 4.0], [5.0, 6.0]]) * 10.0
-#     batch_dim_2 = batch_dim_1 * 2.0
-#     data = torch.stack((batch_dim_1, batch_dim_2), dim=0)
-#     t_s = time()
-#     fake_heatmaps = generate_heatmaps(
-#         keypoints=data,
-#         height=256,
-#         width=256,
-#         output_shape=(64, 64),
-#         normalize=True,
-#         nan_heatmap_mode="zero",
-#     )
-#     t_e = time()
-#     t_ours = t_e - t_s
-#     t_s = time()
-#     data[:, :, 0] *= 64.0 / 256.0  # make it 4 times smaller
-#     data[:, :, 1] *= 64.0 / 256.0  # make it 4 times smaller
-#     kornia_heatmaps = render_gaussian2d(
-#         mean=data.reshape(-1, 2), std=torch.tensor((1.0, 1.0)), size=(64, 64)
-#     )
-#     t_e = time()
-#     t_kornia = t_e - t_s
-#     print(kornia_heatmaps[0, :, :].flatten())
-#     print(fake_heatmaps[0, :, :].flatten())
-#     print((kornia_heatmaps[0, :, :].flatten()).sum())
-#     print((fake_heatmaps[0, :, :].flatten().sum()))
-#     kornia_heatmaps = kornia_heatmaps.reshape(2, 3, 64, 64)
-#     kornia_min_max = (kornia_heatmaps.min(), kornia_heatmaps.max())
-#     print(kornia_min_max)
-#     our_min_max = (fake_heatmaps.min(), fake_heatmaps.max())
-#     print(our_min_max)
-#     data_1 = data.reshape(-1, 2)
-#     data_2 = data_1.reshape(2, 3, 2)
-#     (data == data_2).all()
-#     kornia_heatmaps.shape
-#     fake_keypoints = torch.tensor(3)
+def test_normalized_to_bbox():
+
+    from lightning_pose.data.utils import normalized_to_bbox
+
+    # test when keypoints and bboxes are same size
+    keypoints = torch.tensor([
+        [[0.0, 0.0]],  # xy for 1 keypoint
+        [[1.0, 1.0]],
+        [[0.5, 0.5]],
+    ])
+
+    bboxes = [
+        torch.tensor([0, 0, 100, 200]),  # xyhw
+        torch.tensor([20, 30, 100, 200]),
+    ]
+    for bbox in bboxes:
+        kps = normalized_to_bbox(keypoints.clone(), bbox.unsqueeze(0).repeat([3, 1]))
+        # (0.0, 0.0) should map to top left corner
+        assert kps[0, 0, 0] == bbox[0]
+        assert kps[0, 0, 1] == bbox[1]
+        # (1.0, 1.0) should map to bottom right corner
+        assert kps[1, 0, 0] == bbox[3] + bbox[0]
+        assert kps[1, 0, 1] == bbox[2] + bbox[1]
+        # (0.5, 0.5) should map to top left corner plus half the new height/width
+        assert kps[2, 0, 0] == bbox[3] / 2 + bbox[0]
+        assert kps[2, 0, 1] == bbox[2] / 2 + bbox[1]
+
+    # test when keypoints come from context model and bboxes have extra entries for edges
+    for bbox in bboxes:
+        kps = normalized_to_bbox(keypoints.clone(), bbox.unsqueeze(0).repeat([7, 1]))
+        # (0.0, 0.0) should map to top left corner
+        assert kps[0, 0, 0] == bbox[0]
+        assert kps[0, 0, 1] == bbox[1]
+        # (1.0, 1.0) should map to bottom right corner
+        assert kps[1, 0, 0] == bbox[3] + bbox[0]
+        assert kps[1, 0, 1] == bbox[2] + bbox[1]
+        # (0.5, 0.5) should map to top left corner plus half the new height/width
+        assert kps[2, 0, 0] == bbox[3] / 2 + bbox[0]
+        assert kps[2, 0, 1] == bbox[2] / 2 + bbox[1]
+
+
+def test_convert_bbox_coords(heatmap_data_module, multiview_heatmap_data_module):
+
+    from lightning_pose.data.utils import convert_bbox_coords
+
+    # -------------------------------------
+    # test on single view dataset
+    # -------------------------------------
+
+    # params
+    x_crop = 25
+    y_crop = 40
+
+    # get training batch
+    batch_dict = next(iter(heatmap_data_module.train_dataloader()))
+    orig_converted = convert_bbox_coords(batch_dict, batch_dict['keypoints'])
+    old_image_dims = [batch_dict['images'].size(-2), batch_dict['images'].size(-1)]
+    old_bbox = batch_dict["bbox"]
+    x_pix = x_crop * old_bbox[:, 3] / old_image_dims[1]
+    y_pix = y_crop * old_bbox[:, 2] / old_image_dims[0]
+
+    # create a new batch with smaller & cropped images
+    new_dict = batch_dict
+    new_dict['images'] = new_dict['images'][:, :, y_crop:-y_crop, x_crop:-x_crop]
+    new_dict['bbox'][:, 0] = new_dict['bbox'][:, 0] + x_pix
+    new_dict['bbox'][:, 1] = new_dict['bbox'][:, 1] + y_pix
+    new_dict['bbox'][:, 2] = new_dict['bbox'][:, 2] - 2 * y_pix
+    new_dict['bbox'][:, 3] = new_dict['bbox'][:, 3] - 2 * x_pix
+    new_dict['keypoints'][:, 0::2] += x_crop  # keypoints x,y shifted in image
+    new_dict['keypoints'][:, 1::2] += y_crop
+    new_converted = convert_bbox_coords(new_dict, new_dict['keypoints'])
+
+    # orig and new converted coordinates should be the same
+    assert torch.allclose(orig_converted, new_converted, equal_nan=True)
+
+    # -------------------------------------
+    # test on dummy multi view dataset
+    # -------------------------------------
+    batch_dict = {
+        "images": torch.tensor(np.random.randn(2, 2, 3, 10, 10)),  # batch, views, RGB, h, w
+        "predicted_keypoints": torch.tensor([
+            [0.0, 0.0, 0.0, 0.0],  # xy, xy (2 keypoints
+            [10.0, 10.0, 10.0, 10.0],
+        ]),
+        "bbox": torch.tensor([
+            [5.0, 6.0, 100.0, 101.0, 10.0, 11.0, 102.0, 103.0],  # xyhw x 2
+            [0.0, 0.0, 123.0, 124.0, 0.0, 0.0, 3.0, 4.0],
+        ]),
+        "num_views": torch.tensor([2, 2]),
+    }
+    converted = convert_bbox_coords(batch_dict, batch_dict["predicted_keypoints"])
+    assert converted[0, 0] == batch_dict["bbox"][0, 0]
+    assert converted[0, 1] == batch_dict["bbox"][0, 1]
+    assert converted[0, 2] == batch_dict["bbox"][0, 4]
+    assert converted[0, 3] == batch_dict["bbox"][0, 5]
+    assert converted[1, 0] == batch_dict["bbox"][1, 3]
+    assert converted[1, 1] == batch_dict["bbox"][1, 2]
+    assert converted[1, 2] == batch_dict["bbox"][1, 7]
+    assert converted[1, 3] == batch_dict["bbox"][1, 6]
+
+    # -------------------------------------
+    # test on dummy multi view context dataset
+    # -------------------------------------
+    batch_dict = {
+        "images": torch.tensor(np.random.randn(2, 2, 3, 10, 10)),  # batch, views, RGB, h, w
+        "predicted_keypoints": torch.tensor([
+            [0.0, 0.0, 0.0, 0.0],  # xy, xy (2 keypoints)
+            [10.0, 10.0, 10.0, 10.0],
+        ]),
+        "bbox": torch.tensor([
+            [1.0, 2.0, 100.0, 101.0, 10.0, 11.0, 102.0, 103.0],  # context, will be removed
+            [1.0, 2.0, 100.0, 101.0, 10.0, 11.0, 102.0, 103.0],  # context, will be removed
+            [5.0, 6.0, 100.0, 101.0, 10.0, 11.0, 102.0, 103.0],  # xyhw x 2
+            [0.0, 0.0, 123.0, 124.0, 0.0, 0.0, 3.0, 4.0],
+            [1.0, 2.0, 100.0, 101.0, 10.0, 11.0, 102.0, 103.0],  # context, will be removed
+            [1.0, 2.0, 100.0, 101.0, 10.0, 11.0, 102.0, 103.0],  # context, will be removed
+        ]),
+        "num_views": torch.tensor([2, 2, 2, 2, 2, 2]),
+    }
+    converted = convert_bbox_coords(batch_dict, batch_dict["predicted_keypoints"])
+    assert converted[0, 0] == batch_dict["bbox"][2, 0]
+    assert converted[0, 1] == batch_dict["bbox"][2, 1]
+    assert converted[0, 2] == batch_dict["bbox"][2, 4]
+    assert converted[0, 3] == batch_dict["bbox"][2, 5]
+    assert converted[1, 0] == batch_dict["bbox"][3, 3]
+    assert converted[1, 1] == batch_dict["bbox"][3, 2]
+    assert converted[1, 2] == batch_dict["bbox"][3, 7]
+    assert converted[1, 3] == batch_dict["bbox"][3, 6]
+
+    # -------------------------------------
+    # test error on multi view dataset
+    # -------------------------------------
+    # get training batch
+    batch_dict = next(iter(multiview_heatmap_data_module.train_dataloader()))
+    # change number of views for one batch element
+    batch_dict["num_views"][0] = 16
+    # make sure code complains when batch elements have different numbers of views
+    with pytest.raises(ValueError):
+        convert_bbox_coords(batch_dict, batch_dict['keypoints'])
